Hedge Ratio, Hedging Effectiveness, and Optimal Hedging Weight

Fits a bivariate asymmetric DCC (aDCC) model with GJR–GARCH(1,1) margins and reports: (i) the time-varying hedge ratio summary (mean/min/max), (ii) hedging effectiveness (HE) value based on conditional variances, and (iii) the optimal portfolio weight of the hedging asset averaged over time (weights clipped to [0,1]).

Usage

hedge_effectiveness_dcc(
  hedged,
  hedge,
  arma_order = c(1, 0),
  garch_order = c(1, 1),
  univariate_dist = "norm",
  dcc_dist = "mvnorm",
  dcc_model = "aDCC"
)

Arguments

hedged

Numeric vector or zoo/xts series of returns for the asset being hedged (asset 1).

hedge

Numeric vector or zoo/xts series of returns for the hedging asset (asset 2).

arma_order

Integer length-2; ARMA order in the univariate mean (default c(1, 0)).

garch_order

Integer length-2; GARCH order in the univariate variance (default c(1, 1)).

univariate_dist

Character; univariate error distribution for rugarch::ugarchspec() (default "norm").

dcc_dist

Character; DCC distribution for rmgarch::dccspec() (default "mvnorm").

dcc_model

Character; DCC model type (default "aDCC").

Value

A one-row data.frame with:

beta_mean

Mean hedge ratio \(\bar{\beta}\)

beta_min

Minimum hedge ratio

beta_max

Maximum hedge ratio

HE

Single conditional hedging effectiveness value

OPW

Mean optimal portfolio weight of the hedging asset (clipped)

Details

Let \(\Sigma_t\) be the conditional covariance matrix from the DCC fit with elements \(\Sigma_{11,t} = \sigma^2_{1,t}\), \(\Sigma_{22,t} = \sigma^2_{2,t}\), and \(\Sigma_{12,t}\). The time-varying hedge ratio is $$\beta_t = \Sigma_{12,t} / \Sigma_{22,t}.$$

The conditional hedging effectiveness (HE) is reported as a single value: $$HE \;=\; 1 - \frac{\overline{\operatorname{Var}_t(y_1 - \beta_t y_2)}}% {\overline{\sigma^2_{1,t}}} \;=\; 1 - \frac{\overline{\sigma^2_{1,t} - \Sigma_{12,t}^2/\sigma^2_{2,t}}} {\overline{\sigma^2_{1,t}}},$$ where the bar denotes the time average.

The optimal portfolio weight of the hedged/spot asset \(S\) versus the hedging asset \(F\) follows: $$w_{SF,t} \;=\; \frac{\sigma^2_{F,t} - \Sigma_{SF,t}} {\sigma^2_{S,t} - 2\Sigma_{SF,t} + \sigma^2_{F,t}},$$ clipped to \([0,1]\). The function returns OPW as the mean weight of the hedging asset, i.e. \(\overline{1 - w_{SF,t}}\).

References

Basher, S. A., & Sadorsky, P. (2016). Hedging emerging market stock prices with oil, gold, VIX, and bonds: A comparison between DCC, ADCC and GO-GARCH. Energy Economics, 54, 235–247.

See also

Examples

if (FALSE) { # \dontrun{
# Hedge stock (SP) with gold (GLD)
hedge_effectiveness_dcc(hedgedata$SP, hedgedata$GLD)
} # }